requirements.txt
fastapi[standard]
uvicorn[standard]
pydantic-settings
alembic
SQLAlchemy
psycopg2
bcrypt
pyjwt
pycryptodome
requests